MXS-3340 Make database and user creation identical
Make database and user creation identical in the case of
regular nodes and Galera nodes.

- Output is not piped to /dev/null.
- Database is first dropped and then created.
- RESET MASTER is executed.

The Galera script still waits for the Galera node to become
ready before creating things. That needs to be lifted out from
that script after which the Galera script can be dropped, and
the same script be used in all cases.